backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
uvloop==0.19.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
pandas==2.1.4
python-dateutil==2.8.2
httpx==0.26.0
google-generativeai==0.5.4
pdfplumber==0.10.3
